refactor(config, error):패키지 추가
