+++ src/test/java/lotto/domain/model/lotto/LottoTest.java
@@ -3,7 +3,7 @@ package lotto.domain.model.lotto;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import lotto.ErrorMessage;
+import lotto.error.ErrorMessage;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
